JWT agregado y validar rol

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\PruebaController;
 use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
@@ -67,7 +68,9 @@ Route::get('/token', function () {
     Route::get('/api/user/{id}', [UserController::class,'getByIdUser'])->middleware('validateIdUser');
     Route::post('/api/user', [UserController::class,'addUser'])->middleware('validateInputs');
 
-    Route::put('/api/user/{id}', [UserController::class,'updateUser'])->middleware('validateIdUser', 'validateInputs');
-    Route::delete('/api/user/{id}', [UserController::class,'deleteUser'])->middleware('validateIdUser');
+    Route::put('/api/user/{id}', [UserController::class,'updateUser'])->middleware('validateJWT','validateIdUser', 'validateInputs');
+    Route::delete('/api/user/{id}', [UserController::class,'deleteUser'])->middleware('validateJWT','isAdminRole','validateIdUser');
 
+    // LOGIN
+    Route::post('/api/auth', [AuthController::class, 'login'])->middleware('validateLogin');
